+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 更新日志
 
+24w08a
+---
+- CHANGE: 同步更新logger模块，与golang-temp项目定义的开发规范保持一致
+- ADD: 新增日志翻转功能
+
 v1.1.1
 ---
 - CHANGE: 修改部分代码，与golang-temp项目定义的开发规范保持一致